+++ src/cmds.io.cc
@@ -62,7 +62,7 @@
 #if !CURL_AT_LEAST_VERSION(7, 80, 0)
 extern "C"
 {
-    const char* curl_url_strerror(CURLUcode error);
+const char* curl_url_strerror(CURLUcode error);
 }
 #endif
 
@@ -211,9 +211,9 @@ com_save_to(exec_context& ec,
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file name")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);
@@ -774,9 +774,9 @@ com_open(exec_context& ec, std::string cmdline, std::vector<std::string>& args)
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file names")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);
@@ -1287,9 +1287,9 @@ com_xopen(exec_context& ec, std::string cmdline, std::vector<std::string>& args)
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file names")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);
@@ -1330,9 +1330,9 @@ com_close(exec_context& ec, std::string cmdline, std::vector<std::string>& args)
             auto split_err = split_args_res.unwrapErr();
             auto um = lnav::console::user_message::error(
                           "unable to parse file name")
-                          .with_reason(split_err.te_msg)
+                          .with_reason(split_err.se_error.te_msg)
                           .with_snippet(lnav::console::snippet::from(
-                              SRC, lexer.to_attr_line(split_err)))
+                              SRC, lexer.to_attr_line(split_err.se_error)))
                           .move();
 
             return Err(um);
@@ -1628,9 +1628,9 @@ com_redirect_to(exec_context& ec,
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file name")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);