+++ src/lnav_commands.cc
@@ -362,9 +362,9 @@ com_set_file_timezone(exec_context& ec,
             auto split_err = split_res.unwrapErr();
             auto um = lnav::console::user_message::error(
                           "unable to parse arguments")
-                          .with_reason(split_err.te_msg)
+                          .with_reason(split_err.se_error.te_msg)
                           .with_snippet(lnav::console::snippet::from(
-                              SRC, lexer.to_attr_line(split_err)))
+                              SRC, lexer.to_attr_line(split_err.se_error)))
                           .move();
 
             return Err(um);
@@ -1799,9 +1799,9 @@ com_file_visibility(exec_context& ec,
             auto split_err = split_args_res.unwrapErr();
             auto um = lnav::console::user_message::error(
                           "unable to parse file name")
-                          .with_reason(split_err.te_msg)
+                          .with_reason(split_err.se_error.te_msg)
                           .with_snippet(lnav::console::snippet::from(
-                              SRC, lexer.to_attr_line(split_err)))
+                              SRC, lexer.to_attr_line(split_err.se_error)))
                           .move();
 
             return Err(um);
@@ -2827,9 +2827,9 @@ com_cd(exec_context& ec, std::string cmdline, std::vector<std::string>& args)
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file name")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);
@@ -3710,9 +3710,9 @@ com_prompt(exec_context& ec,
             auto split_err = split_args_res.unwrapErr();
             auto um
                 = lnav::console::user_message::error("unable to parse prompt")
-                      .with_reason(split_err.te_msg)
+                      .with_reason(split_err.se_error.te_msg)
                       .with_snippet(lnav::console::snippet::from(
-                          SRC, lexer.to_attr_line(split_err)))
+                          SRC, lexer.to_attr_line(split_err.se_error)))
                       .move();
 
             return Err(um);