+++ src/sql_commands.cc
@@ -119,9 +119,9 @@ sql_cmd_read(exec_context& ec,
         auto split_err = split_args_res.unwrapErr();
         auto um
             = lnav::console::user_message::error("unable to parse file name")
-                  .with_reason(split_err.te_msg)
+                  .with_reason(split_err.se_error.te_msg)
                   .with_snippet(lnav::console::snippet::from(
-                      SRC, lexer.to_attr_line(split_err)))
+                      SRC, lexer.to_attr_line(split_err.se_error)))
                   .move();
 
         return Err(um);