+++ test/test_text_file.sh
@@ -164,6 +164,6 @@ run_cap_test ${lnav_test} -nN -c ':config /tuning/piper/max-size 128'
 
 cat ${test_dir}/textfile_plain.0 | \
   run_cap_test ${lnav_test} -n -d /tmp/lnav-text.err \
-    -c ';SELECT filepath FROM lnav_file'
+    -c ';SELECT filepath FROM lnav_file ORDER BY filepath DESC'
 
 #####